+++ src/main/java/nl/jusx/pycharm/lineprofiler/render/FunctionProfileInlayRenderer.java
@@ -4,9 +4,10 @@ import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.EditorCustomElementRenderer;
 import com.intellij.openapi.editor.Inlay;
 import com.intellij.openapi.editor.markup.TextAttributes;
+import jViridis.ColorMap;
 import nl.jusx.pycharm.lineprofiler.profile.FunctionProfile;
 import nl.jusx.pycharm.lineprofiler.profile.Profile;
-import jViridis.ColorMap;
+import nl.jusx.pycharm.lineprofiler.settings.SettingsState;
 import org.jetbrains.annotations.NotNull;
 
 import java.awt.*;
@@ -67,7 +68,7 @@ public class FunctionProfileInlayRenderer implements EditorCustomElementRenderer
         currentX = tableAlignment.align(currentX, "timeInFunction");
 
         g.drawString(
-                String.format("%.0f %s",
+                String.format("%d %s",
                         functionProfile.getTotalTime(),
                         profile.getUnitLong()),
                 currentX,
@@ -110,7 +111,7 @@ public class FunctionProfileInlayRenderer implements EditorCustomElementRenderer
     private void paintColormap(Graphics g, int x, int y, int width, int height) {
         Graphics2D g2 = (Graphics2D) g;
 
-        ColorMap cm = ColorMap.getInstance(ColorMap.VIRIDIS);
+        ColorMap cm = ColorMap.getInstance(SettingsState.getInstance().getColorMap().getIdentifier());
         Color startColor, endColor;
 
         startColor = cm.getColor(0f);