+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2022-02-16
+### Added
+- More colormaps. Different colormaps can now be selected in the plugin settings.
+
+### Changed
+- Made the max table alignment column configurable
+
+### Fixed
+- Use longs for 'time' and 'hits' instead of ints to prevent NumberFormatExceptions
+
 ## [1.4.0] - 2021-12-01
 ### Changed
 - Removed deprecated code
@@ -31,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial version with a profile executor
 
 [Unreleased]: https://gitlab.com/line-profiler-pycharm/line-profiler-pycharm-plugin/tree/dev
+[1.5.0]: https://gitlab.com/line-profiler-pycharm/line-profiler-pycharm-plugin/-/tags/1.5.0
 [1.4.0]: https://gitlab.com/line-profiler-pycharm/line-profiler-pycharm-plugin/-/tags/1.4.0
 [1.3.0]: https://gitlab.com/line-profiler-pycharm/line-profiler-pycharm-plugin/-/tags/1.3.0
 [1.2.0]: https://gitlab.com/line-profiler-pycharm/line-profiler-pycharm-plugin/-/tags/1.2.0